More scaladoc updates; Incorporate review feedback ... EC needs re factored, unsafe access to state in progress still
